+++ src/types/types.ts
@@ -77,6 +77,34 @@ export enum TransactionPayloadVariants {
 }
 
 /**
+ * The inner payload type to support turbo transactions and all future transaction types.
+ * {@link https://github.com/aptos-labs/aptos-core/blob/main/types/src/transaction/mod.rs#L478}
+ */
+export enum TransactionInnerPayloadVariants {
+  V1 = 0,
+}
+
+/**
+ * Executable types for transactions, which can be either a script or an entry function.
+ *
+ * Empty is reserved for Multisig voting transactions, which do not have an executable payload.
+ * {@link https://github.com/aptos-labs/aptos-core/blob/main/types/src/transaction/mod.rs#L685}
+ */
+export enum TransactionExecutableVariants {
+  Script = 0,
+  EntryFunction = 1,
+  Empty = 2,
+}
+
+/**
+ * Variants of transaction extra configurations, which can include additional settings or parameters.
+ * {@link https://github.com/aptos-labs/aptos-core/blob/main/types/src/transaction/mod.rs#L737}
+ */
+export enum TransactionExtraConfigVariants {
+  V1 = 0,
+}
+
+/**
  * Variants of transactions used in the system.
  * {@link https://github.com/aptos-labs/aptos-core/blob/main/types/src/transaction/mod.rs#L440}
  */